feat(portal): save token after login
