+++ vortex-app/vortex-app-portal/src/components/Access/Authenticate.tsx
@@ -1,10 +1,10 @@
 import { Spin } from 'antd'
-import React, { useEffect, ReactNode } from 'react'
+import React, { useEffect, ReactNode, useCallback } from 'react'
 import { useAuth0 } from '@auth0/auth0-react'
 import { useNavigate } from 'react-router-dom'
 import { filter, get } from 'lodash'
 import { useGetUserAuthDetail, useGetUserRole } from '@/hooks/user'
-import { getOrg } from '@/utils/helpers/token'
+import { getOrg, storeToken } from '@/utils/helpers/token'
 import { ENV } from '@/constant'
 import { useAppStore } from '@/stores/app.store'
 import type { AuthUser } from '@/stores/type'
@@ -14,12 +14,17 @@ interface AuthenticateProps {
 }
 
 const Authenticate = ({ children }: AuthenticateProps) => {
-  const { isLoading, isAuthenticated, user } = useAuth0()
+  const { isLoading, isAuthenticated, user, getAccessTokenSilently } = useAuth0()
   const { currentAuth0User, setCurrentAuth0User, setUser, setRoleList } = useAppStore()
   const navigate = useNavigate()
 
   const { data: userData } = useGetUserAuthDetail()
   const { data: roleData } = useGetUserRole()
+
+  const saveToken = useCallback(async () => {
+    const res = await getAccessTokenSilently()
+    storeToken(res)
+  }, [getAccessTokenSilently])
 
   useEffect(() => {
     const userDetail = userData?.data
@@ -39,10 +44,13 @@ const Authenticate = ({ children }: AuthenticateProps) => {
       const org = getOrg() || ENV.AUTH0_MGMT_ORG_ID
       navigate(`${org}/login`)
     }
+    if (isAuthenticated) {
+      saveToken()
+    }
     if (isAuthenticated && !currentAuth0User && user) {
       setCurrentAuth0User(user as AuthUser)
     }
-  }, [isAuthenticated, isLoading, user])
+  }, [isAuthenticated, isLoading, user, saveToken])
 
   if (isLoading || !isAuthenticated) {
     return <Spin />
